+++ catscripts/Entertainment/Video_games/Video_game_consoles/Microsoft/Xbox360.sh
@@ -2,15 +2,23 @@
 
 KEYWORDS_XBOX360="X(|-)box(| )360"
 
-egrep -i "$KEYWORDS_XBOX360" newpages.txt >> Xbox360.txt
-
-X360=$(stat --print=%s Xbox360.txt)
-
-if [ $X360 -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Xbox360.txt"
-  export CATNAME="Xbox 360"
-  $CATEGORIZE
-fi
 
-rm Xbox360.txt
+  debug_start "Xbox 360"
+
+  X360=$(egrep -i "$KEYWORDS_XBOX360" newpages.txt)
+
+  if [ "$X360" != "" ];
+  then
+    printf "%s" "$X360" > Xbox360.txt
+    export CATFILE="Xbox360.txt"
+    export CATNAME="Xbox 360"
+    $CATEGORIZE
+    rm Xbox360.txt
+    unset X360
+  fi
+
+  debug_end "Xbox 360"
+
+fi
