+++ catscripts/Entertainment/Video_games/Video_games_by_series/Fire_Emblem_series/FireEmblem.sh
@@ -1,15 +1,25 @@
 #!/bin/bash
 
-egrep -i 'Fire(| )Emblem' newpages.txt >> FireEmblem.txt
+KEYWORDS_FIREEMBLEM="Fire(| )Emblem"
 
-EMBLEM=$(stat --print=%s FireEmblem.txt)
-
-if [ $EMBLEM -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="FireEmblem.txt"
-  export CATNAME="Fire Emblem series"
-  $CATEGORIZE
-fi
 
-rm FireEmblem.txt
+  debug_start "Fire Emblem series"
+
+  EMBLEM=$(egrep -i "$KEYWORDS_FIREEMBLEM" newpages.txt)
+
+  if [ "$EMBLEM" != "" ];
+  then
+    printf "%s" "$EMBLEM" > FireEmblem.txt
+    export CATFILE="FireEmblem.txt"
+    export CATNAME="Fire Emblem series"
+    $CATEGORIZE
+    rm FireEmblem.txt
+    unset EMBLEM
+  fi
+
+  debug_end "Fire Emblem series"
+
+fi
